+++ service/src/main/java/com/ericsson/eiffel/remrem/generate/controller/RemremGenerateController.java
@@ -1,5 +1,5 @@
 /*
-    Copyright 2017 Ericsson AB.
+    Copyright 2018 Ericsson AB.
     For a full list of individual contributors, please see the commit history.
     Licensed under the Apache License, Version 2.0 (the "License");
     you may not use this file except in compliance with the License.
@@ -17,12 +17,17 @@ package com.ericsson.eiffel.remrem.generate.controller;
 import com.ericsson.eiffel.remrem.generate.constants.RemremGenerateServiceConstants;
 import com.ericsson.eiffel.remrem.protocol.MsgService;
 import com.ericsson.eiffel.remrem.shared.VersionService;
+import com.google.gson.Gson;
+import com.google.gson.GsonBuilder;
 import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
 import com.google.gson.JsonParser;
 
+import io.swagger.annotations.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
+import org.springframework.http.RequestEntity;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -30,14 +35,18 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
+import springfox.documentation.annotations.ApiIgnore;
 
-import java.util.Collection;
 import java.util.List;
 import java.util.Map;
 
 @RestController
 @RequestMapping("/*")
+@Api(value = "REMReM Generate Service", description = "REST API for generating Eiffel messages")
 public class RemremGenerateController {
+
+    // regular expression that exclude "swagger-ui.html" from request parameter
+    private static final String REGEX = ":^(?!swagger-ui.html).*$";
 
     @Autowired
     private List<MsgService> msgServices;
@@ -49,17 +58,23 @@ public class RemremGenerateController {
      * <p>
      * <p>
      * Parameters:
-     * mp - The message protocol , which tells us which service to invoke.
+     * msgProtocol - The message protocol, which tells us which service to invoke.
      * msgType - The type of message that needs to be generated.
      * bodyJson - The content of the message which is used in creating the event details.
      * <p>
      * Returns:
      * The event information as a json element
      */
-    @RequestMapping(value = "/{mp}", method = RequestMethod.POST)
-    public ResponseEntity<?> generate(@PathVariable String mp, @RequestParam("msgType") String msgType,
-                                      @RequestBody JsonObject bodyJson) {
-        MsgService msgService = getMessageService(mp);
+    @ApiOperation(value = "To generate eiffel event based on the message protocol", response = String.class)
+    @ApiResponses(value = { @ApiResponse(code = 200, message = "Event sent successfully"),
+            @ApiResponse(code = 400, message = "Malformed JSON"),
+            @ApiResponse(code = 500, message = "Internal server error"),
+            @ApiResponse(code = 503, message = "Message protocol is invalid") })
+    @RequestMapping(value = "/{mp" + REGEX + "}", method = RequestMethod.POST)
+    public ResponseEntity<?> generate(@ApiParam(value = "message protocol", required = true) @PathVariable("mp") final String msgProtocol,
+                                      @ApiParam(value = "message type", required = true) @RequestParam("msgType") final String msgType,
+                                      @ApiParam(value = "JSON message", required = true) @RequestBody final JsonObject bodyJson) {
+        MsgService msgService = getMessageService(msgProtocol);
         String response;
         try {
             if (msgService != null) {
@@ -72,7 +87,7 @@ public class RemremGenerateController {
                 }
             } else {
                 return new ResponseEntity<>(parser.parse(RemremGenerateServiceConstants.NO_SERVICE_ERROR),
-                                            HttpStatus.SERVICE_UNAVAILABLE);
+                        HttpStatus.SERVICE_UNAVAILABLE);
             }
         } catch (Exception e) {
             return new ResponseEntity<>(null, HttpStatus.INTERNAL_SERVER_ERROR);
@@ -80,48 +95,107 @@ public class RemremGenerateController {
     }
 
     /**
-     * this method is used to display the versions of generate and all loaded protocols.
+     * Used to display the versions of generate and all loaded protocols.
      *
-     * @return this method returns a json with version details.
+     * @return json with version details.
      */
+    @ApiOperation(value = "To get versions of generate and all loaded protocols", response = String.class)
     @RequestMapping(value = "/versions", method = RequestMethod.GET)
     public JsonElement getVersions() {
         Map<String, Map<String, String>> versions = new VersionService().getMessagingVersions();
         return parser.parse(versions.toString());
     }
-    
+
     /**
-     * this method returns available Eiffel event types as listed in EiffelEventType enum.
+     * Returns available Eiffel event types as listed in EiffelEventType enum.
      *
      * @return string collection with event types.
      */
+    @ApiOperation(value = "To get available eiffel event types based on the message protocol", response = String.class)
+    @ApiResponses(value = { @ApiResponse(code = 200, message = "Event  types got successfully"),
+            @ApiResponse(code = 500, message = "Internal server error"),
+            @ApiResponse(code = 503, message = "Message protocol is invalid") })
     @RequestMapping(value = "/event_types/{mp}", method = RequestMethod.GET)
-    public ResponseEntity<Collection<String>> getEventTypes(@PathVariable("mp") String mp) {
-    	MsgService msgService = getMessageService(mp);    	
-    	return new ResponseEntity<Collection<String>>(msgService.getSupportedEventTypes(), HttpStatus.OK);
+    public ResponseEntity<?> getEventTypes(@ApiParam(value = "message protocol", required = true) @PathVariable("mp") final String msgProtocol,
+                                           @ApiIgnore final RequestEntity requestEntity) {
+        MsgService msgService = getMessageService(msgProtocol);
+        try {
+            if (msgService != null) {
+                return presentResponse(msgService.getSupportedEventTypes(), HttpStatus.OK, requestEntity);
+            } else {
+                return presentResponse(parser.parse(RemremGenerateServiceConstants.NO_SERVICE_ERROR),
+                        HttpStatus.SERVICE_UNAVAILABLE, requestEntity);
+            }
+        } catch (Exception e) {
+            return presentResponse(parser.parse(RemremGenerateServiceConstants.INTERNAL_SERVER_ERROR),
+                    HttpStatus.INTERNAL_SERVER_ERROR, requestEntity);
+        }
     }
-    
+
     /**
      * Returns an eiffel event template matching the type specified in the path.
      *
      * @return json containing eiffel event template.
      */
+    @ApiOperation(value = "To get eiffel event template of specified event type", response = String.class)
+    @ApiResponses(value = { @ApiResponse(code = 200, message = "Template got successfully"),
+            @ApiResponse(code = 400, message = "Requested template is not available"),
+            @ApiResponse(code = 500, message = "Internal server error"),
+            @ApiResponse(code = 503, message = "Message protocol is invalid") })
     @RequestMapping(value = "/template/{type}/{mp}", method = RequestMethod.GET)
-    public ResponseEntity<JsonElement> getEventTypeTemplate(@PathVariable("type") String type, @PathVariable("mp") String mp) {
-        MsgService msgService = getMessageService(mp);
-        JsonElement template = msgService.getEventTemplate(type);
-        if(template != null)
-            return new ResponseEntity<JsonElement>(template, HttpStatus.OK);
-        else
-            return new ResponseEntity<JsonElement>(template, HttpStatus.NOT_FOUND);
+    public ResponseEntity<?> getEventTypeTemplate(@ApiParam(value = "message type", required = true) @PathVariable("type") final String msgType,
+                                                  @ApiParam(value = "message protocol", required = true) @PathVariable("mp") final String msgProtocol,
+                                                  @ApiIgnore final RequestEntity requestEntity) {
+        MsgService msgService = getMessageService(msgProtocol);
+        try {
+            if (msgService != null) {
+                JsonElement template = msgService.getEventTemplate(msgType);
+                if (template != null) {
+                    return presentResponse(template, HttpStatus.OK, requestEntity);
+                } else {
+                    return presentResponse(parser.parse(RemremGenerateServiceConstants.NO_TEMPLATE_ERROR) , HttpStatus.NOT_FOUND, requestEntity);
+                }
+            } else {
+                return presentResponse(parser.parse(RemremGenerateServiceConstants.NO_SERVICE_ERROR), HttpStatus.SERVICE_UNAVAILABLE, requestEntity);
+            }
+        } catch (Exception e) {
+            return presentResponse(parser.parse(RemremGenerateServiceConstants.INTERNAL_SERVER_ERROR), HttpStatus.INTERNAL_SERVER_ERROR, requestEntity);
+        }
     }
 
-    private MsgService getMessageService(String messageProtocol) {
+    private MsgService getMessageService(final String messageProtocol) {
         for (MsgService service : msgServices) {
             if (service.getServiceName().equals(messageProtocol)) {
                 return service;
             }
         }
         return null;
+    }
+
+    /**
+     * To display pretty formatted json in browser
+     * @param rawJson json content
+     * @return html formatted json string
+     */
+    private String buildHtmlReturnString(final String rawJson) {
+        final String htmlHead = "<!DOCTYPE html><html><body><pre>";
+        final String htmlTail = "</pre></body></html>";
+        return htmlHead + rawJson + htmlTail ;
+    }
+
+    /**
+     * To display response in browser or application
+     * @param message json content
+     * @param status  response code for the HTTP request
+     * @param requestEntity entity of the HTTP request
+     * @return entity to present response in browser or application
+     */
+    private ResponseEntity<?> presentResponse(final Object message, final HttpStatus status, final RequestEntity requestEntity) {
+        if (requestEntity.getHeaders().getAccept().contains(MediaType.TEXT_HTML)) {
+            Gson gson = new GsonBuilder().setPrettyPrinting().create();
+            return new ResponseEntity<>(buildHtmlReturnString(gson.toJson(message)), status);
+        } else {
+            return new ResponseEntity<>(message, status);
+        }
     }
 }